Move impl to helix-view-aggregator module
