+++ helix-view-aggregator/src/test/java/org/apache/helix/view/dataprovider/TestViewClusterDataCache.java
@@ -1,4 +1,23 @@
-package org.apache.helix.common.caches;
+package org.apache.helix.view.dataprovider;
+
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
 import java.util.HashSet;
 import java.util.Set;
@@ -11,7 +30,7 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 
-public class TestBasicClusterDataCache extends ZkTestBase {
+public class TestViewClusterDataCache extends ZkTestBase {
   private static final int NUM_INSTANCE = 5;
 
   private final String CLUSTER_NAME = CLUSTER_PREFIX + "_" + getShortClassName();
@@ -31,7 +50,7 @@ public class TestBasicClusterDataCache extends ZkTestBase {
 
   @Test
   public void testCacheUpdate() {
-    BasicClusterDataCache cache = new BasicClusterDataCache(CLUSTER_NAME);
+    ViewClusterDataCache cache = new ViewClusterDataCache(CLUSTER_NAME);
     ZKHelixDataAccessor accessor = new ZKHelixDataAccessor(CLUSTER_NAME, _baseAccessor);
     Assert.assertTrue(cache.updateCache(accessor));
     Assert.assertEquals(cache.getLiveInstances().size(), NUM_INSTANCE);
